refactor cls test dir
